send functions for guessing from utilities to factorization
